requirements.txt
streamlit==1.8.1
numpy
pandas
Pillow
